added a FreeCAD file for the raw water intake
Currently there's just a single sketch for the weir.

This weir should fall 20 cm onto a screen of an intake, which will likely be made out of a standard 200-liter HDPE food-grade barrel that's been cut in half length-wise.

As these barrels have a diameter of 580 - 590 mm, I think we should make:

 1. The lower height of the weir (where the water falls) 20 cm + 600 mm = 26 cm.Or maybe we should round-up to 30 cm?

 2. The width of the weir (where the water comes-out) should be < 580 mm. Let's just make it 0.5 m.

For redundancy, we want to have two outflows in the weir falling onto two distinct intakes. I decided to just sketch one 0.5 m outflow with a 0.25 m buffer on either side.

My plan is to mirror the whole thing across the axis, as it should be symmetrical. That should simplify the CAD file.